LPS-56986 Rename packages for consistency
