+++ modules/sdk/ant-bnd/test/unit/com/liferay/ant/bnd/bower/dependencies/bower.bad.version.json
