+++ modules/sdk/ant-bnd/test/unit/com/liferay/ant/bnd/spring/dependencies/META-INF/spring/context.dependencies
